型をDubleに変更

+++ scala-study/Distance.scala
@@ -1,6 +1,6 @@
 object Distance extends App {
-    val kilometersPerHours: Float = 11
-    val minutes: Float = 100
+    val kilometersPerHours: Double = 11
+    val minutes: Double = 100
     val distance = kilometersPerHours * minutes / 60
     println(s"走った距離は${distance}キロメートルです")
 }
